+++ programs/mango-v4/tests/test_basic.rs
@@ -39,7 +39,7 @@ async fn test_basic() -> Result<(), TransportError> {
 
     let account = send_tx(
         solana,
-        CreateAccountInstruction {
+        AccountCreateInstruction {
             account_num: 0,
             group,
             owner,
@@ -179,7 +179,7 @@ async fn test_basic() -> Result<(), TransportError> {
     // close account
     send_tx(
         solana,
-        CloseAccountInstruction {
+        AccountCloseInstruction {
             group,
             account,
             owner,